+++ src/ebird/dataset/data/models/county.py
@@ -26,6 +26,10 @@ class County(models.Model):
         null=True, auto_now_add=True, help_text=_("When was the record created.")
     )
 
+    modified = models.DateTimeField(
+        null=True, auto_now=True, help_text=_("When was the record updated.")
+    )
+
     def __repr__(self) -> str:
         return str(self.code)