+++ src/ebird/dataset/data/models/observation.py
@@ -4,6 +4,47 @@ from django.utils.translation import gettext_lazy as _
 
 
 class Observation(models.Model):
+    class BreedingCode(models.TextChoices):
+        NEST_WITH_YOUNG = "NY", _("Nest with young")
+        NEST_WIH_EGGS = "NE", _("Nest containing eggs")
+        FECAL_SAC = "FS", _("Adult carrying a faecal sac")
+        FEEDING_YOUNG = "FY", _("Adult feeding young")
+        CARRYING_FOOD = "CF", _("Adult carrying food")
+        RECENTLY_FLEDGED = "FL", _("Recently fledged young")
+        OCCUPIED_NEST = "ON", _("Occupied nest")
+        USED_NEST = "UN", _("Used nest")
+        DISTRACTION_DISPLAY = "DD", _("Distraction display")
+        NEST_BUILDING = "NB", _("Nest-building")
+        CARRYING_MATERIAL = "CN", _("Carrying nest material")
+        BROOD_PATCH = "PE", _("Brood patch")
+        APPARENT_NEST = "B", _("Woodpecker or Wren nest-building")
+        AGITATED_BEHAVIOUR = "A", _("Agitated behaviour")
+        NEST_VISITED = "N", _("Nest-site visited")
+        COURTSHIP = "C", _("Courtship, display, or copulation")
+        TERRITORY = "T", _("Territorial defense")
+        PAIR = "P", _("Pair in suitable habitat")
+        MULTIPLE_SINGING = "M", _("Multiple singing males")
+        MULTIPLE_DAYS = "S7", _("Male singing for multiple days")
+        SINGING = "S", _("Singing male")
+        HABITAT = "H", _("Adult in suitable habitat")
+        FLYOVER = "FO", _("Fly-over")
+
+    class BreedingCategory(models.TextChoices):
+        OBSERVED = "C1", _("Bird observed")
+        POSSIBLE = "C2", _("Breeding possible")
+        PROBABLE = "C3", _("Breeding probable")
+        CONFIRMED = "C4", _("Breeding confirmed")
+
+    class Age(models.TextChoices):
+        ADULT = "adult", _("Adult")
+        IMMATURE = "immature", _("Immature")
+        JUVENILE = "juvenile", _("Juvenile")
+
+    class Sex(models.TextChoices):
+        FEMALE = "femail", _("Female")
+        MALE = "male", _("Male")
+        UNKNOWN = "unknown", _("Sex unknown")
+
     class Meta:
         verbose_name = _("observation")
         verbose_name_plural = _("observations")
@@ -13,10 +54,6 @@ class Observation(models.Model):
         primary_key=True,
         verbose_name=_("identifier"),
         help_text=_("The identifier for the observation."),
-    )
-
-    created = models.DateTimeField(
-        null=True, auto_now_add=True, help_text=_("When was the record created.")
     )
 
     edited = models.DateTimeField(
@@ -109,20 +146,26 @@ class Observation(models.Model):
         help_text=_("The number of birds seen."),
     )
 
-    breeding_code = models.TextField(
+    breeding_code = models.CharField(
+        max_length=2,
         blank=True,
+        db_index=True,
         verbose_name=_("breeding code"),
         help_text=_("eBird code identifying the breeding status."),
     )
 
-    breeding_category = models.TextField(
+    breeding_category = models.CharField(
+        max_length=2,
         blank=True,
+        db_index=True,
         verbose_name=_("breeding category"),
         help_text=_("eBird code identifying the breeding category."),
     )
 
-    behavior_code = models.TextField(
+    behavior_code = models.CharField(
+        max_length=2,
         blank=True,
+        db_index=True,
         verbose_name=_("behaviour code"),
         help_text=_("eBird code identifying the behaviour."),
     )
@@ -166,6 +209,14 @@ class Observation(models.Model):
         blank=True,
         verbose_name=_("URN"),
         help_text=_("The globally unique identifier for the observation."),
+    )
+
+    created = models.DateTimeField(
+        auto_now_add=True, help_text=_("When was the record created."),
+    )
+
+    modified = models.DateTimeField(
+        auto_now=True, help_text=_("When was the record updated.")
     )
 
     def __repr__(self) -> str: