+++ src/ebird/dataset/data/models/species.py
@@ -2,41 +2,35 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 
-SPECIES_CATEGORY = {
-    "species": _("Species"),
-    "sub-species": _("Sub-species"),
-    "hybrid": _("Hybrid"),
-    "intergrade": _("Intergrade"),
-    "spuh": _("Genus"),
-    "slash": _("Species group"),
-    "domestic": _("Domestic"),
-    "form": _("Form"),
-}
-
-EXOTIC_CODE = {
-    "": "",  # NATIVE
-    "N": _("Naturalized"),
-    "P": _("Provisional"),
-    "X": _("Escapee"),
-}
-
-
 class Species(models.Model):
+    class Category(models.TextChoices):
+        SPECIES = "species", _("Species")
+        SLASH = "slash", _("Species pairs")
+        SUBSPECIES = "issf", _("Subspecies")
+        DOMESTIC = "domestic", _("Domestic species")
+        HYBRID = "hybrid", _("Hybrids")
+        FORM = "form", _("Species forms")
+        SPUH = "spuh", _("Unidentified species")
+        INTERGRADE = "intergrade", _("Intergrades")
+
+    class Exotic(models.TextChoices):
+        NATURALIZED = "N", _("Naturalized")
+        PROVISIONAL = "P", _("Provisional")
+        ESCAPEE = "X", _("Escapee")
+
     class Meta:
         verbose_name = _("species")
         verbose_name_plural = _("species")
 
-    created = models.DateTimeField(
-        auto_now_add=True,
-        help_text=_("When was the record created."),
-    )
-
     order = models.IntegerField(
+        db_index=True,
         verbose_name=_("taxonomic order"),
         help_text=_("The position in the eBird/Clements taxonomic order."),
     )
 
     category = models.TextField(
+        choices=Category,
+        db_index=True,
         verbose_name=_("category"),
         help_text=_("The category from the eBird/Clements taxonomy."),
     )
@@ -74,8 +68,17 @@ class Species(models.Model):
 
     exotic_code = models.TextField(
         blank=True,
+        db_index=True,
         verbose_name=_("exotic code"),
         help_text=_("The code used if the species is non-native."),
+    )
+
+    created = models.DateTimeField(
+        auto_now_add=True, help_text=_("When was the record created."),
+    )
+
+    modified = models.DateTimeField(
+        auto_now=True, help_text=_("When was the record updated.")
     )
 
     def __repr__(self) -> str: